fix capsule doc
