--- tutorials/first.py
@@ -1,14 +1,0 @@
-"""
-Your first example in DGL
-=========================
-
-TODO: either a pagerank or SSSP example
-"""
-
-###############################################################################
-# Create a DGLGraph
-# -----------------
-#
-# To start with, let's first import dgl
-
-import dgl
